cast int64_t in asserts

+++ aten/src/ATen/native/Linear.cpp
@@ -5,22 +5,22 @@
 namespace at { namespace native {
 
 Tensor bilinear(const Tensor& input1, const Tensor& input2, const Tensor& weight, const Tensor& bias) {
-  AT_ASSERT(input1.dim() == input2.dim(), "bilinear(): input dimensions do not match: got %d and %d",
-            input1.dim(), input2.dim());
+  AT_ASSERT(input1.dim() == input2.dim(), "bilinear(): input dimensions do not match: got %lld and %lld",
+            (long long)input1.dim(), (long long)input2.dim());
   for (int64_t i = 0; i < input1.dim() - 1; i++) {
     AT_ASSERT(input1.size(i) == input2.size(i),
-              "bilinear(): input batch dimensions do not match at dim %d: got %d and %d",
-              i, input1.size(i), input2.size(i));
+              "bilinear(): input batch dimensions do not match at dim %lld: got %lld and %lld",
+              (long long)i, (long long)input1.size(i), (long long)input2.size(i));
   }
   AT_ASSERT(input1.size(input1.dim() - 1) == weight.size(1),
-            "bilinear(): input1 size does not match weight size: got %d but expected %d",
-            input1.size(input1.dim() - 1), weight.size(1));
+            "bilinear(): input1 size does not match weight size: got %lld but expected %lld",
+            (long long)input1.size(input1.dim() - 1), (long long)weight.size(1));
   AT_ASSERT(input2.size(input2.dim() - 1) == weight.size(2),
-            "bilinear(): input2 size does not match weight size: got %d but expected %d",
-            input2.size(input2.dim() - 1), weight.size(2));
+            "bilinear(): input2 size does not match weight size: got %lld but expected %lld",
+            (long long)input2.size(input2.dim() - 1), (long long)weight.size(2));
   AT_ASSERT(bias.defined() && bias.size(0) == weight.size(0),
-            "bilinear(): bias size does not match weight size: got %d but expected %d",
-            bias.size(0), weight.size(0));
+            "bilinear(): bias size does not match weight size: got %lld but expected %lld",
+            (long long)bias.size(0), (long long)weight.size(0));
 
   auto b_input1 = input1.unsqueeze(-2).unsqueeze(-2);
   auto b_input2 = input2.unsqueeze(-2).unsqueeze(-1);
